backend/requirements.txt
Django==5.2.8
django-cors-headers==4.9.0
django-jazzmin==3.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.3
gunicorn==23.0.0
psycopg2-binary
python-dotenv
dj-database-url
whitenoise
requests